Primeiro refatoramento do código do game. Separando as classes em módulos externos.
